+++ lib/src/authsignal_flutter_platform.dart
@@ -54,7 +54,7 @@ abstract class AuthsignalFlutterPlatform extends PlatformInterface {
     String? displayName,
     bool useAutoRegister = false,
     bool ignorePasskeyAlreadyExistsError = false,
-    bool syncCredentials = true,
+    bool syncCredentials = false,
   }) {
     throw UnimplementedError('passkeySignUp() has not been implemented.');
   }
@@ -64,7 +64,7 @@ abstract class AuthsignalFlutterPlatform extends PlatformInterface {
     String? token,
     bool autofill = false,
     bool preferImmediatelyAvailableCredentials = true,
-    bool syncCredentials = true,
+    bool syncCredentials = false,
   }) {
     throw UnimplementedError('passkeySignIn() has not been implemented.');
   }